location module implemented with authorization
