+++ src/modules/location/location.routes.js
@@ -1,17 +1,22 @@
 import {Router} from 'express';
 import {celebrate, Segments} from 'celebrate';
-import *  as LocationController from './location.contrllers';
+import { authLocal, authJwt } from '../../services/auth.services';
+import *  as LocationController from './location.controllers';
 
 const routes = new Router();
 
-routes.post('/create', LocationController.createLocation);
+//create a location for a user in the system
+routes.post('/',authJwt, LocationController.createLocation);
 
-routes.get('/get/:id',LocationController.getLocation);
+//read a location related to a user
+routes.get('/user/',authJwt,LocationController.getUserLocation);
 
-routes.patch('/update/:id',LocationController.updateLocation);
+//update a location related to a user
+routes.patch('/user/',authJwt,LocationController.updateUserLocation);
 
-routes.delete('/delete/:id',LocationController.deleteLocation);
+//delete a location related to a user
+routes.delete('/user/',authJwt,LocationController.deleteUserLocation);
 
-routes.get('/getUser/:id',LocationController.getUserLocation);
+routes.get('/getUser/:id',authJwt,LocationController.getUserLocation);
 
 export default routes;
